+++ Cargo.toml
@@ -74,12 +74,7 @@ members = [
 
 [features]
 default = ["jitdump", "wasmtime/wat", "wasmtime/parallel-compilation"]
-lightbeam = [
-    "wasmtime-environ/lightbeam",
-    "wasmtime-jit/lightbeam",
-    "wasmtime-wast/lightbeam",
-    "wasmtime/lightbeam",
-]
+lightbeam = ["wasmtime/lightbeam"]
 jitdump = ["wasmtime/jitdump"]
 vtune = ["wasmtime/vtune"]
