+++ crates/cranelift/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "wasmtime-cranelift"
+version = "0.19.0"
+authors = ["The Wasmtime Project Developers"]
+description = "Integration between Cranelift and Wasmtime"
+license = "Apache-2.0 WITH LLVM-exception"
+repository = "https://github.com/bytecodealliance/wasmtime"
+documentation = "https://docs.rs/wasmtime-cranelift/"
+categories = ["wasm"]
+keywords = ["webassembly", "wasm"]
+readme = "README.md"
+edition = "2018"
+
+[dependencies]
+wasmtime-environ = { path = "../environ", version = "0.19.0" }
+cranelift-wasm = { path = "../../cranelift/wasm", version = "0.66.0" }
+cranelift-codegen = { path = "../../cranelift/codegen", version = "0.66.0" }
+cranelift-frontend = { path = "../../cranelift/frontend", version = "0.66.0" }
+cranelift-entity = { path = "../../cranelift/entity", version = "0.66.0" }
