+++ crates/cranelift/README.md
@@ -1,0 +1,4 @@
+# `wasmtime-cranelfit`
+
+This crate provides an implementation of the `Compiler` trait which is
+connected to Cranelift.
