+++ crates/environ/Cargo.toml
@@ -15,10 +15,8 @@ edition = "2018"
 anyhow = "1.0"
 cranelift-codegen = { path = "../../cranelift/codegen", version = "0.66.0", features = ["enable-serde"] }
 cranelift-entity = { path = "../../cranelift/entity", version = "0.66.0", features = ["enable-serde"] }
-cranelift-frontend = { path = "../../cranelift/frontend", version = "0.66.0" }
 cranelift-wasm = { path = "../../cranelift/wasm", version = "0.66.0", features = ["enable-serde"] }
 wasmparser = "0.59.0"
-lightbeam = { path = "../lightbeam", optional = true, version = "0.19.0" }
 indexmap = { version = "1.0.2", features = ["serde-1"] }
 thiserror = "1.0.4"
 serde = { version = "1.0.94", features = ["derive"] }
