+++ crates/lightbeam/wasmtime/README.md
@@ -1,0 +1,4 @@
+# `wasmtime-lightbeam`
+
+This crate provides an implementation of the `Compiler` trait which is
+connected to Lightbeam.
